move type into core
