+++ packages/mobx-state-tree/src/core/index.ts
@@ -1,4 +1,5 @@
 export * from "./node"
+export * from "./type"
 export { createActionInvoker } from "./action"
 export * from "./json-patch"
 export * from "./mst-operations"
