+++ packages/mobx-state-tree/src/core/mst-operations.ts
@@ -649,5 +649,4 @@ import { IObservableArray, ObservableMap } from "mobx"
 import { INode, getStateTreeNode, IStateTreeNode, isStateTreeNode } from "./node"
 import { IJsonPatch, splitJsonPath } from "./json-patch"
 import { asArray, EMPTY_OBJECT, fail, IDisposer } from "../utils"
-import { ISnapshottable, IType } from "../types/type"
-import { isType } from "../types/type-flags"
+import { ISnapshottable, IType, isType } from "./type"
