+++ packages/mobx-state-tree/src/core/node/object-node.ts
@@ -158,7 +158,7 @@ import { reaction } from "mobx"
 import { ScalarNode } from "./scalar-node"
 import { INode, isStateTreeNode, getStateTreeNode } from "./node-utils"
 import { IJsonPatch, IReversibleJsonPatch, splitJsonPath, splitPatch } from "../json-patch"
-import { IType } from "../../types/type"
+import { IType } from "../type"
 import { IDisposer, extend, noop, fail, registerEventHandler, addReadOnlyProp } from "../../utils"
 import { walk } from "../mst-operations"
 import { IMiddlewareHandler, createActionInvoker } from "../action"
