+++ packages/mobx-state-tree/src/core/type/type-flags.ts
@@ -1,15 +1,15 @@
 import { IType, IComplexType } from "./type"
-import { CoreType } from "./primitives"
-import { IExtendedObservableMap } from "./complex-types/map"
-import { ModelType } from "./complex-types/model"
-import { IdentifierType } from "./utility-types/identifier"
-import { ReferenceType } from "./utility-types/reference"
-import { Refinement } from "./utility-types/refinement"
-import { Union } from "./utility-types/union"
-import { OptionalValue } from "./utility-types/optional"
-import { Frozen } from "./utility-types/frozen"
-import { Late } from "./utility-types/late"
-import { Literal } from "./utility-types/literal"
+import { CoreType } from "../../types/primitives"
+import { IExtendedObservableMap } from "../../types/complex-types/map"
+import { ModelType } from "../../types/complex-types/model"
+import { IdentifierType } from "../../types/utility-types/identifier"
+import { ReferenceType } from "../../types/utility-types/reference"
+import { Refinement } from "../../types/utility-types/refinement"
+import { Union } from "../../types/utility-types/union"
+import { OptionalValue } from "../../types/utility-types/optional"
+import { Frozen } from "../../types/utility-types/frozen"
+import { Late } from "../../types/utility-types/late"
+import { Literal } from "../../types/utility-types/literal"
 import { IObservableArray } from "mobx"
 
 export enum TypeFlags {
