+++ packages/mobx-state-tree/src/core/type/type.ts
@@ -197,8 +197,8 @@ export abstract class Type<S, T> extends ComplexType<S, T> implements IType<S, T
     }
 }
 
-import { EMPTY_ARRAY, fail, isMutable } from "../utils"
-import { isStateTreeNode, getStateTreeNode } from "../core/node/node-utils"
+import { EMPTY_ARRAY, fail, isMutable } from "../../utils"
+import { isStateTreeNode, getStateTreeNode } from "../node/node-utils"
 import {
     IContext,
     IValidationResult,
@@ -206,6 +206,6 @@ import {
     typeCheckFailure,
     typeCheckSuccess
 } from "./type-checker"
-import { INode, IStateTreeNode } from "../core/node/node-utils"
-import { IJsonPatch } from "../core/json-patch"
-import { getType } from "../core/mst-operations"
+import { INode, IStateTreeNode } from "../node/node-utils"
+import { IJsonPatch } from "../json-patch"
+import { getType } from "../mst-operations"
